+++ NEWS.md
@@ -1,4 +1,6 @@
 # mitre 0.3.0
+ 
+* Added CAR
 
 # mitre 0.2.0