+++ NEWS.md
@@ -1,3 +1,5 @@
+# mitre (development version)
+
 # mitre 0.3.0
  
 * Added CAR